requirements.txt
django>=4.2.0
django-tailwind>=3.8.0
djangorestframework==3.15.2
